Break the interface into a separate file
