+++ src/comm.ts
@@ -10,42 +10,9 @@ import {
 
 import { createKernelMessage } from './kernel';
 
+import { IComm, ICommManager, ICommInfo } from './icomm';
+
 import * as utils from './utils';
-
-
-/**
- * A manager for Comm objects.
- */
-export
-interface ICommManager {
-  startNewComm(targetName: string, data: any, commId?: string): Promise<IComm>;
-  connectToComm(targetName: string, commId: string): Promise<IComm>;
-  registerTarget(targetName: string, cb: (comm: IComm, data: any) => any): void;
-  listRunningComms(targetName: string): Promise<ICommInfo[]>;
-}
-
-
-/**
- * A client side Comm interface.
- */
-export
-interface IComm {
-  commId: string;
-  targetName: string;
-  send(data: any): void;
-  close(data?: any): void;
-  onClose: (data?: any) => void;
-  onMsg: (data: any) => void;
-}
-
-
-/**
- * Contents of a `comm_info` message response.
- */
-export
-interface ICommInfo {
-  // TBD
-}
 
 
 /*
@@ -215,6 +182,7 @@ class CommManager implements ICommManager {
 /**
  * Comm channel handler.
  */
+export
 class Comm implements IComm {
 
   /**
